Split testtool
Make rebase
