+++ src/wazuh_modules/vulnerability_scanner/testtool/CMakeLists.txt
@@ -4,3 +4,4 @@ add_subdirectory(scanner)
 add_subdirectory(databaseFeedManager)
 add_subdirectory(wazuhDBQuery)
 add_subdirectory(rocksDBQuery)
+add_subdirectory(versionMatcher)
