+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/CMakeLists.txt
@@ -1,16 +1,16 @@
 cmake_minimum_required(VERSION 3.12.4)
-project(database_feed_manager_testtool)
+project(version_matcher_testtool)
 
 set(CMAKE_CXX_FLAGS_DEBUG "-g -fsanitize=address,leak,undefined")
 
 link_directories(${SRC_FOLDER}/external/curl/lib/.libs/)
 
-file(GLOB DATABASE_FEED_MANAGER_TESTTOOL_SRC
+file(GLOB VERSION_MATCHER_TESTTOOL_SRC
     "*.cpp"
     )
 
 add_executable(${PROJECT_NAME}
-    ${DATABASE_FEED_MANAGER_TESTTOOL_SRC}
+    ${VERSION_MATCHER_TESTTOOL_SRC}
     )
 
 target_link_libraries(${PROJECT_NAME}
@@ -22,4 +22,5 @@ target_link_libraries(${PROJECT_NAME}
         rocksdb
         urlrequest
         curl
+        scan_orchestrator
 )
