+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/argsParser.hpp
@@ -1,7 +1,7 @@
 /*
  * Wazuh cmdLineParser
  * Copyright (C) 2015, Wazuh Inc.
- * October 26, 2023.
+ * Dec 20, 2023.
  *
  * This program is free software; you can redistribute it
  * and/or modify it under the terms of the GNU General Public
@@ -33,7 +33,8 @@ public:
         : m_configurationFilePath {paramValueOf(argc, argv, "-c")}
         , m_waitTime {paramValueOf(argc, argv, "-s", std::make_pair(false, "0"))}
         , m_inputFiles {splitActions(paramValueOf(argc, argv, "-i", std::make_pair(false, "")))}
-        , m_formatInputFiles {paramValueOf(argc, argv, "-f", std::make_pair(false, "offsets"))}
+        , m_formatInputFiles {paramValueOf(argc, argv, "-f", std::make_pair(false, "offset"))}
+        , m_checkVersionMatcher {paramValueOf(argc, argv, "-v", std::make_pair(false, "no"))}
     {
     }
 
@@ -44,6 +45,15 @@ public:
     const std::string& getConfigurationFilePath() const
     {
         return m_configurationFilePath;
+    }
+
+    /**
+     * @brief Gets the configuration verifier.
+     * @return Configuration version matcher.
+     */
+    const std::string& getConfigurationVersionMatcher() const
+    {
+        return m_checkVersionMatcher;
     }
 
     /**
@@ -78,17 +88,19 @@ public:
      */
     static void showHelp()
     {
-        std::cout << "\nUsage: DatabaseFeedManager <option(s)>\n"
+        std::cout << "\nUsage: versionMatcher <option(s)>\n"
                   << "Options:\n"
                   << "\t-h \t\t\tShow this help message\n"
                   << "\t-c CONFIG_FILE\tSpecifies the configuration file.\n"
                   << "\t-s WAIT_TIME\tSpecifies the wait before exit.\n"
                   << "\t-i INPUT_FILES\tSpecifies the input files.\n"
                   << "\t-f [offset | raw]\tSpecifies the format of the input files.\n"
+                  << "\t-v [yes | no]\tRun a verification against candidates with versionMatcher algorithms.\n"
                   << "\nExample:"
-                  << "\n\t./DatabaseFeedManager -c config.json\n"
-                  << "\n\t./DatabaseFeedManager -c config.json -s 10\n"
-                  << "\n\t./DatabaseFeedManager -c config.json -f offset -i file1.json,file2.json\n"
+                  << "\n\t./versionMatcher -c config.json\n"
+                  << "\n\t./versionMatcher -c config.json -s 10\n"
+                  << "\n\t./versionMatcher -c config.json -f offset -i file1.json,file2.json\n"
+                  << "\n\t./versionMatcher -c config.json -f offset -i file1.json,file2.json -v yes\n"
                   << std::endl;
     }
 
@@ -136,6 +148,7 @@ private:
     const std::string m_waitTime;
     const std::vector<std::string> m_inputFiles;
     const std::string m_formatInputFiles;
+    const std::string m_checkVersionMatcher;
 };
 
 #endif // _CMD_ARGS_PARSER_HPP_
