+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/config.json
@@ -1,0 +1,3 @@
+{
+    "topicName" : "test"
+}
