+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/file2.json
@@ -1,0 +1,35 @@
+{
+    "data": [
+        {
+            "offset": 5,
+            "type": "create",
+            "version": 1,
+            "context": "vulnerabilities",
+            "resource": "TID-0001",
+            "payload": {
+                "target": "windows",
+                "source": {
+                    "vendor": "^VMware",
+                    "product": "^VMware Workstation$",
+                    "version": ""
+                },
+                "translation": [
+                    {
+                        "vendor": "vmware",
+                        "product": "workstation",
+                        "version": ""
+                    },
+                    {
+                        "vendor": "vmware",
+                        "product": "workstation_pro (version >= 12.0.0)",
+                        "version": ""
+                    }
+                ],
+                "action": [
+                    "replace_vendor",
+                    "replace_product_if_matches"
+                ]
+            }
+        }
+    ]
+}
