+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/file_offset1.json
@@ -1,0 +1,122 @@
+{
+    "data": [
+        {
+            "offset": 1,
+            "type": "create",
+            "version": 1,
+            "context": "vulnerabilities",
+            "resource": "CVE-1999-1234",
+            "payload": {
+                "containers": {
+                    "adp": [
+                        {
+                            "affected": [
+                                {
+                                    "defaultStatus": "unknown",
+                                    "product": "bash",
+                                    "vendor" : "debian",
+                                    "platforms": [
+                                        "upstream"
+                                    ],
+                                    "versions": [
+                                        {
+                                            "lessThan": "98.0.4758.102",
+                                            "status": "affected",
+                                            "version": "0",
+                                            "versionType": "custom"
+                                        }
+                                    ]
+                                }
+                            ],
+                            "descriptions": [
+                                {
+                                    "lang": "en",
+                                    "value": "The /etc/profile.d/60alias.sh script in the Mandriva bash package for Bash\n2.05b, 3.0, 3.2, 3.2.48, and 4.0 enables the --show-control-chars option in\nLS_OPTIONS, which allows local users to send escape sequences to terminal\nemulators, or hide the existence of a file, via a crafted filename."
+                                }
+                            ],
+                            "providerMetadata": {
+                                "orgId": "00000000-0000-4000-A000-000000000000",
+                                "shortName" : "debian"
+                            },
+                            "references": [
+                                {
+                                    "url": "https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2010-0002"
+                                }
+                            ]
+                        }
+                    ],
+                    "cna": {
+                        "affected": [
+                            {
+                                "cpes": [
+                                    "cpe:2.3:a:gnu:bash:2.05:b:*:*:*:*:*:*"
+                                ],
+                                "defaultStatus": "unaffected",
+                                "product": "bash b",
+                                "vendor": "gnu",
+                                "versions": [
+                                    {
+                                        "status": "affected",
+                                        "version": "2.05"
+                                    }
+                                ]
+                            }
+                        ],
+                        "descriptions": [
+                            {
+                                "lang": "en",
+                                "value": "The /etc/profile.d/60alias.sh script in the Mandriva bash package for Bash 2.05b, 3.0, 3.2, 3.2.48, and 4.0 enables the --show-control-chars option in LS_OPTIONS, which allows local users to send escape sequences to terminal emulators, or hide the existence of a file, via a crafted filename."
+                            }
+                        ],
+                        "metrics": [
+                            {
+                                "cvssV2_0": {
+                                    "accessComplexity": "LOW",
+                                    "accessVector": "LOCAL",
+                                    "authentication": "NONE",
+                                    "availabilityImpact": "PARTIAL",
+                                    "baseScore": 2.1,
+                                    "confidentialityImpact": "NONE",
+                                    "integrityImpact": "NONE",
+                                    "vectorString": "AV:L/AC:L/Au:N/C:N/I:N/A:P",
+                                    "version": "2.0"
+                                },
+                                "format": "CVSS"
+                            }
+                        ],
+                        "problemTypes": [
+                            {
+                                "descriptions": [
+                                    {
+                                        "description": "CWE-20",
+                                        "lang": "en"
+                                    }
+                                ]
+                            }
+                        ],
+                        "providerMetadata": {
+                            "orgId": "00000000-0000-4000-A000-000000000000",
+                            "shortName": "nvd"
+                        },
+                        "references": [
+                            {
+                                "name": "https://qa.mandriva.com/show_bug.cgi?id=56882",
+                                "url": "https://qa.mandriva.com/show_bug.cgi?id=56882"
+                            }
+                        ]
+                    }
+                },
+                "cveMetadata": {
+                    "assignerOrgId": "00000000-0000-4000-A000-000000000000",
+                    "assignerShortName": "@redhat.com",
+                    "cveId": "CVE-1999-1234",
+                    "datePublished": "2010-01-14T18:30:00.000Z",
+                    "dateUpdated": "2011-08-08T04:00:00.000Z",
+                    "state": "PUBLISHED"
+                },
+                "dataType": "CVE_RECORD",
+                "dataVersion": "5.0"
+            }
+        }
+    ]
+}
