+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/file_raw1.json
@@ -1,0 +1,1 @@
+{"name":"CVE-2023-30518","offset":198730,"version":1,"inserted_at":"2023-08-04T14:00:34.936827Z","payload":{"containers":{"cna":{"descriptions":[{"lang":"en","value":"A missing permission check in Jenkins Thycotic Secret Server Plugin 1.0.2 and earlier allows attackers with Overall/Read permission to enumerate credentials IDs of credentials stored in Jenkins."}],"references":[{"url":"https://www.jenkins.io/security/advisory/2023-04-12/#SECURITY-2837"},{"url":"http://www.openwall.com/lists/oss-security/2023/04/13/3"}]}},"cveMetadata":{"cveId":"CVE-2023-30518","datePublished":"2023-04-12T17:05:09.301Z","state":"PUBLISHED"},"dataType":"CVE_RECORD","dataVersion":"5.0"}}
