+++ src/wazuh_modules/vulnerability_scanner/testtool/versionMatcher/message.json
@@ -1,0 +1,5 @@
+{
+    "paths": [
+        "file1.json"
+    ]
+}
